+++ packages/terra-clinical-action-header/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Update to React 16
 
+### Added
+* prop `keepCloseButton` to keep the close button in small viewports
+
 1.6.0 - (December 9, 2017)
 ----------
 ### Added